Add CI workflow + fix cross-platform bugs found while validating it locally
CI (.github/workflows/build.yml): Linux + Windows (MinGW cross-compile,
no MSYS2) + 4 Android NDK ABIs (arm64-v8a, armeabi-v7a, x86_64, x86) =
8 artifacts, matching the established tar_repacker pattern. Windows is
cross-compiled from Ubuntu via the mingw-w64 apt package rather than an
MSYS2 runner, to avoid the DLL-search-path interference that's bitten
this codebase's other projects before.

Before writing the workflow, ran both the Linux and the exact MinGW
cross-compile command locally (apt install mingw-w64) to catch real
breakage instead of debugging blind in CI. That surfaced four genuine
bugs:

- S_ISLNK/S_ISSOCK/S_IFIFO and O_LARGEFILE aren't defined by MinGW's CRT
  headers. Centralized fallback definitions in win_pread.h (previously
  only f2fs_extract.cpp had a partial/local fix; f2fs_metadata.cpp's
  typeChar() needed S_ISLNK/S_ISSOCK too and didn't have it at all).
- applyTimestamps() declared a local struct named __utimbuf64 (double
  underscore) but cast to _utimbuf64* (single underscore) when calling
  _utime64() -- a different, incomplete type. Now uses the system-
  provided struct directly instead of redeclaring it, removing the
  typo class entirely.
- %zu in several log format strings isn't supported by MinGW's printf
  (msvcrt-based by default) -- switched to %llu with explicit
  (unsigned long long) casts throughout.
- libzstd_static's INTERFACE_INCLUDE_DIRECTORIES pointed into the build
  tree without a BUILD_INTERFACE generator-expression wrapper, which
  fails CMake's Generate-step validation as soon as any install(TARGETS)
  exists in the project (zstd's own CMakeLists.txt installs
  libzstd_static by default, triggering the check). This was silently
  masked on Linux by a leftover system libzstd-dev install from earlier
  manual testing in this environment -- confirmed by temporarily
  removing liblz4-dev/libzstd-dev/liblzo2-dev system-wide and rebuilding
  clean, which is also how the missing zstd.h include path was caught
  in the first place.

Also added CMAKE_POLICY_VERSION_MINIMUM=3.5 before FetchContent -- zstd's
bundled CMakeLists.txt declares cmake_minimum_required(VERSION 2.8.12),
which CMake 4.x refuses outright without this escape hatch. Same root
cause as the CMake 4.x issues flagged in this codebase's CI history
before; addressed proactively here rather than waiting for a runner
image bump to break it.

Verified after all fixes: full regression suite (compress round-trip,
plain-file round-trip, real odm.img LZ4 sepolicy extraction, SELinux/
capability xattr) passes on both a from-scratch Linux build and the
from-scratch MinGW cross-build, with zero system compression packages
installed in either case.

+++ src/f2fs_metadata.cpp
@@ -31,6 +31,7 @@
 #include <cstring>
 #include <filesystem>
 #include <sys/stat.h>
+#include "win_pread.h"  // POSIX stat-macro shims (S_ISLNK etc.) on Windows; passthrough on POSIX
 
 namespace fs = std::filesystem;
 
@@ -191,7 +192,7 @@ bool MetadataWriter::writeJSON(const std::string& path) const
     }
 
     fputs("{\n  \"format\": \"f2fs_extract_metadata\",\n  \"version\": 1,\n", f);
-    fprintf(f, "  \"entry_count\": %zu,\n", entries_.size());
+    fprintf(f, "  \"entry_count\": %llu,\n", (unsigned long long)entries_.size());
     fputs("  \"entries\": [\n", f);
 
     for (size_t idx = 0; idx < entries_.size(); ++idx) {
